+++ tool/tool.cabal
@@ -25,6 +25,7 @@ source-repository head
 
 library
   exposed-modules:
+      GenerateYaml
       HaskellTypes
       TH
   other-modules:
